\name{bolten98}
\alias{bolten98}
\alias{bolten98raw}
\docType{data}
\title{Loggerhead turtle data from Bolten 1998}
\description{
  Mitochondrial haplotype samples from Bolten 1998
}
\usage{data(bolten98)}
\format{
  A data frame with 17 observations on the following 7 variables.
  \describe{
    \item{\code{NWFL}}{haplotype sample from northwest Florida}
    \item{\code{SOFL}}{ditto, south Florida}
    \item{\code{NEFL.NC}}{northeast Florida to North Carolina}
    \item{\code{Mexico}}{Mexico}
    \item{\code{Greece}}{Greece}
    \item{\code{Brazil}}{Brazil}
    \item{\code{feed}}{feeding grounds}
  }
}
\source{Bolten et al. 1998}
\examples{
  data(bolten98)
}
\keyword{datasets}
