\name{hwdiag.check}
\alias{hwdiag.check}
\title{check Heidelberger and Welch diagnostic results}
\description{
  Runs the Heidelberger and Welch tests (\code{heidel.diag}) implemented
  in coda and tests that all results are OK
}
\usage{
hwdiag.check(x, verbose = FALSE)
}
\arguments{
  \item{x}{the results of a model fit?}
  \item{verbose}{verbose output?}
}
\value{
  logical: true or false
}
\keyword{misc}
