\name{lahanas98}
\title{Green turtle data from Lahanas et al. 1998}
\usage{data(lahanas98)}
\alias{lahanas98}
\alias{lahanas98raw}
\format{(lahanas98) a mixstock.data object;
  (lahanas98raw) a text table with 10 columns (9 rookeries + feeding ground) and
  21 rows (haplotypes)}
\description{This data set gives counts of different mitochondrial
  haplotypes sampled in different rookeries throughout the Atlantic and
  Mediterranean (Florida (FL), Mexico (MEXI), Costa Rica (CR), Aves
  Island (AVES), Surinam (SURI),
  Brazil (BRAZ), Ascension Island (ASCE), Africa (AFRI), Cyprus (CYPR)), as well as in a pooled
  feeding-ground population.}
\source{Lahanas et al. 1998}
\keyword{datasets}
